fix: move fastify server to separate file, use top level await
