+++ .eslintrc
@@ -5,7 +5,7 @@
     "es2021": true
   },
   "parserOptions": {
-    "ecmaVersion": 2021,
+    "ecmaVersion": 2022,
     "sourceType": "module"
   },
   "rules": {
